First step in QML GUI implementation
This is a restart of the windowing system from scratch using QQuickView
and qml for the GUI.  The entire code of StelMainView has to change, so
to make things easier I added a StelQuickView class that is eventually
going to be renamed StelMainView.

For the moment all it does is create the main QQuickView and render
the sky into it.  There is no UI or event handlers.