Add a bunch of Q_PROPERTY
This is needed to access the modules from the qml GUI.

I think a lot of code could be avoided by using qt MEMBER instead of
getter and setter for each attribute.  Also we don't need to have a
separate signal for each property: a single 'changed' signal could be
used for all the properties that don't change too often.