Use a position cache for orbit position
It is based on a QMap (Qt implementation of red/black tree).  That way
we can relatively quickly find the nearest cached position for a given
time.

The get method of the PosCache class uses a dt argument to allow en error
range in the returned value.  For the moment I statically set the error
range used to one hour.  I guess we could make this smaller without
changing the performance much.