Use parametric path for planet orbits
This allows to have orbits with arbitrary precision, but it makes the
code really slow.

I added a new StelPainter method to draw parametric path, taking a
std::function as an argument (so that we can use c++ lambda when we call
it).

I removed all the caching from the orbit code, since it was assuming a
fixed size cache, and this does not work anymore.  This is why this
patch makes the code so slow.  I will add a new caching system in a next
commit.